Extract hooks into their own interface
